refactor: use poetry
