+++ acapy_wallet_upgrade/__main__.py
@@ -19,10 +19,10 @@ import msgpack
 import nacl.pwhash
 from urllib.parse import urlparse
 
-from aries_askar_upgrade.db_connection import DbConnection
-from aries_askar_upgrade.sqlite_connection import SqliteConnection
-from aries_askar_upgrade.pg_connection import PgConnection
-from aries_askar_upgrade.error import UpgradeError
+from .db_connection import DbConnection
+from .sqlite_connection import SqliteConnection
+from .pg_connection import PgConnection
+from .error import UpgradeError
 
 
 CHACHAPOLY_KEY_LEN = 32
@@ -579,7 +579,7 @@ async def upgrade(db: DbConnection, master_pw: str):
     print("done")
 
 
-if __name__ == "__main__":
+def main():
     logging.basicConfig(level=logging.WARN)
 
     if len(sys.argv) < 2:
@@ -599,3 +599,7 @@ if __name__ == "__main__":
 
     key = sys.argv[2]  # Faber.Agent372766
     asyncio.get_event_loop().run_until_complete(upgrade(conn, key))
+
+
+if __name__ == "__main__":
+    main()
