+++ pyproject.toml
@@ -1,0 +1,42 @@
+[tool.poetry]
+name = "acapy-wallet-upgrade"
+version = "0.1.0"
+description = "Migrate ACA-Py from Indy SDK to Aries Askar"
+authors = [
+	"Andrew Whitehead <cywolf@gmail.com>",
+	"Adam Burdett <burdettadam@gmail.com>",
+	"Char Howland <char@indicio.tech>",
+	"Daniel Bluhm <dbluhm@pm.me>"
+]
+license = "Apache-2.0"
+classifiers = [
+	"Programming Language :: Python :: 3",
+	"License :: OSI Approved :: Apache Software License",
+	"Operating System :: OS Independent",
+]
+readme = "README.md"
+repository = "https://github.com/indicio-tech/acapy-wallet-upgrade"
+
+[tool.poetry.dependencies]
+python = "^3.6.9"
+aiosqlite = "~=0.17"
+aries-askar = "~=0.2"
+asyncpg = "~=0.22"
+base58 = "~=1.0"
+cbor2 = "~=5.2"
+msgpack = "~=1.0"
+PyNaCl = "~=1.4"
+
+[tool.poetry.dev-dependencies]
+black = "<22.12.0"
+flake8 = "<6.0.0"
+pytest = "<7.2.0"
+pytest-asyncio = "<0.20.3"
+docker-py = "^1.10.6"
+
+[tool.poetry.scripts]
+askar-upgrade = "acapy_wallet_upgrade.__main__:main"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
